Statement classes moved to its own files.
