+++ MathLang.csproj
@@ -45,6 +45,10 @@
     <Compile Include="MathAstNode.cs" />
     <Compile Include="MathLangIntepreter.cs" />
     <Compile Include="Program.cs" />
+    <Compile Include="Tree\Nodes\Statements\ArrayElementAssignmentStatement.cs" />
+    <Compile Include="Tree\Nodes\Statements\DoWhileStatement.cs" />
+    <Compile Include="Tree\Nodes\Statements\ForStatement.cs" />
+    <Compile Include="Tree\Nodes\Statements\IfStatement.cs" />
     <Compile Include="Tree\Nodes\Expressions\AtomExpression.cs" />
     <Compile Include="Tree\Nodes\Declarations\ClassDeclaration.cs" />
     <Compile Include="Tree\Nodes\Expressions\Expression.cs" />
@@ -58,7 +62,10 @@
     <Compile Include="Tree\Nodes\Declarations\FunctionDeclarationParameter.cs" />
     <Compile Include="Tree\Nodes\Enums\ReturnType.cs" />
     <Compile Include="Tree\Nodes\Declarations\VariableDeclaration.cs" />
+    <Compile Include="Tree\Nodes\Statements\ReturnStatement.cs" />
     <Compile Include="Tree\TreeHelper.cs" />
+    <Compile Include="Tree\Nodes\Statements\VariableAssignmentStatement.cs" />
+    <Compile Include="Tree\Nodes\Statements\WhileStatement.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="Input.txt" />
