+++ Tree/Nodes/Statements/IfStatement.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using Antlr.Runtime.Tree;
+using MathLang.Tree.Nodes;
+
+namespace MathLang.Tree
+{
+    public class IfStatement : IStatement
+    {
+        public INode Parent { get; }
+
+        public IfStatement(INode parent)
+        {
+            Parent = parent;
+        }
+
+        public void Construct(CommonTree tree)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
