+++ Tree/Nodes/Statements/VariableAssignmentStatement.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using Antlr.Runtime.Tree;
+using MathLang.Tree.Nodes;
+
+namespace MathLang.Tree
+{
+    public class VariableAssignmentStatement : IStatement
+    {
+        public INode Parent { get; }
+
+        public VariableAssignmentStatement(INode parent)
+        {
+            Parent = parent;
+        }
+        
+        public void Construct(CommonTree tree)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
