+++ Tree/TreeHelper.cs
@@ -114,108 +114,4 @@ namespace MathLang.Tree
             }
         }
     }
-
-    public class ReturnStatement : IStatement
-    {
-        public INode Parent { get; }
-
-        public ReturnStatement(INode parent)
-        {
-            Parent = parent;
-        }
-        
-        public void Construct(CommonTree tree)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class ArrayElementAssignmentStatement : IStatement
-    {
-        public INode Parent { get; }
-
-        public ArrayElementAssignmentStatement(INode parent)
-        {
-            Parent = parent;
-        }
-        
-        public void Construct(CommonTree tree)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class VariableAssignmentStatement : IStatement
-    {
-        public INode Parent { get; }
-
-        public VariableAssignmentStatement(INode parent)
-        {
-            Parent = parent;
-        }
-        
-        public void Construct(CommonTree tree)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class ForStatement : IStatement
-    {
-        public INode Parent { get; }
-
-        public ForStatement(INode parent)
-        {
-            Parent = parent;
-        }
-        public void Construct(CommonTree tree)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class WhileStatement : IStatement
-    {
-        public INode Parent { get; }
-
-        public WhileStatement(INode parent)
-        {
-            Parent = parent;
-        }
-
-        public void Construct(CommonTree tree)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class IfStatement : IStatement
-    {
-        public INode Parent { get; }
-
-        public IfStatement(INode parent)
-        {
-            Parent = parent;
-        }
-
-        public void Construct(CommonTree tree)
-        {
-            throw new NotImplementedException();
-        }
-    }
-    public class DoWhileStatement: IStatement
-    {
-        public INode Parent { get; }
-
-        public DoWhileStatement(INode parent)
-        {
-            Parent = parent;
-        }
-
-        public void Construct(CommonTree tree)
-        {
-            throw new NotImplementedException();
-        }
-
-    }
 }
